[SystemUpdate]: handle Openstore app updates
Change the client API URL to point at the OpenStore, and update the JSON
parsing to match the Openstore's format.
Remove all the code related to the SSO and token download, and remove
the now unneeded dependency on libUbuntuOneAuth. We can bring it back
later, should we ever find the need.

The Openstore support can be improved once the two following bugs are
fixed:
- https://github.com/UbuntuOpenStore/openstore-meta/issues/156
- https://github.com/UbuntuOpenStore/openstore-meta/issues/157