[Security&Privacy] Skip uninstalled apps
This is a workaround for
https://github.com/ubports/ubuntu-touch/issues/385:
when building the list of app permissions, skip those apps for which a
display name cannot be resolved (meaning that they have most likely been
deleted).

This is only a workaround: the application records in the trust-store
should actually be removed, once an application gets uninstalled.