Rejigger ClickAppNotifications page
Turns out there's a switch in ClickAppNotifications.qml which enables
or disables all notification types for the app. It gets automatically
switched when the user disables all notification types. Problem is, it
doesn't show up... it's behind the header!

This commit changes the page around so that the switch shows correctly.
It also makes the page scrollable and fixes a couple of papercuts I saw
from playing around with it. Now when the switch becomes disabled,
all of the notification-specific checkboxes don't display. When the
switch is enabled again, the checkboxes come back. Ahh, satisfying.