Use old icon/grid view on single column layouts
Restore the pre-16.04 layout, with a grid of big icons, when operating
on a single column layout. This is obtained by using a custom positioner
(AdaptiveContainer) which is able to operate both as a Grid and as a
Column.

Rename EntryComponent to EntryComponentList, and bring back the old
EntryComponent as EntryComponentGrid (with only minor changes).

Fixes: https://github.com/ubports/ubuntu-touch/issues/324