Correctly incubate the Updates page
addPageToNextColumn returns an incubating object if the pageStack is in
asynchronous mode (as it is here). We need to wait for the incubated
object to load before we can call check(true) on it.

This fixes the "Check for updates" button in About. It actually
initiates a check for updates now.